+++ src/App.js
@@ -1,15 +1,15 @@
 import React from "react";
-import TopNav from "./components/TopNav";
+import TopNav from "./containers/TopNav";
 import PropTypes from "prop-types";
 import AreaChart from "./components/AreaChart";
-import Comments from "./components/Comments";
+import Comments from "./containers/Comments";
 import DonutChart from "./components/DonutChart";
-import Orders from "./components/Orders";
+import Orders from "./containers/Orders";
 import SideNav from "./components/SideNav";
-import Tasks from "./components/Tasks";
-import TasksPanel from "./components/TasksPanel";
-import Tickets from "./components/Tickets";
-import TransactionsPanel from "./components/TransactionsPanel";
+import Tasks from "./containers/Tasks";
+import TasksPanel from "./containers/TasksPanel";
+import Tickets from "./containers/Tickets";
+import TransactionsPanel from "./containers/TransactionsPanel";
 
 function App(props) {
   return (
@@ -19,7 +19,7 @@ function App(props) {
           className="navbar navbar-inverse navbar-fixed-top"
           role="navigation"
         >
-          <TopNav messages={props.messages} />
+          <TopNav />
           <SideNav />
         </nav>
         <div id="page-wrapper">
@@ -38,18 +38,18 @@ function App(props) {
             </div>
             <div className="row">
               <Comments />
-              <Tasks newTasks={props.newTasks} />
-              <Orders newOrders={props.newOrders} />
-              <Tickets tickets={props.tickets} />
+              <Tasks />
+              <Orders />
+              <Tickets />
             </div>
             <AreaChart />
             <div className="row">
               <DonutChart />
               <div className="col-lg-4">
-                <TasksPanel tasks={props.tasks} />
+                <TasksPanel />
               </div>
               <div className="col-lg-4">
-                <TransactionsPanel orders={props.orders} />
+                <TransactionsPanel />
               </div>
             </div>
           </div>
