+++ src/containers/TopNav.js
@@ -1,0 +1,10 @@
+import { connect } from "react-redux";
+import TopNav from "../components/TopNav";
+
+function mapStateToProps(state) {
+  return {
+    messages: state.messages
+  };
+}
+
+export default connect(mapStateToProps)(TopNav);
